requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
Flask-WTF==1.2.1
email-validator==2.1.0
PyPDF2==3.0.1